requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.21.0
pyyaml>=6.0
tqdm>=4.65.0
opencv-python>=4.7.0
scikit-learn>=1.3.0
pytorch-fid>=0.3.0
accelerate>=0.20.0
peft>=0.4.0
